# openapi.yml
openapi: 3.0.0
info:
  title: SIB Tasks
  description: >-
    This API is made to fulfill the assignment from PT. Arkatama Multisolusindo
    in FWD SIB program.
  version: 1.0.0
servers:
  - url: http://{{base_url}}
...
paths:
...
  /tasks/{task_id}:
    get:
      tags:
        - Task
      summary: Task
      parameters:
        - name: joins[]
          in: query
          schema:
            type: string
          description: Return task with its other relation. Task has a category relation.
          example: category
        - name: task_id
          in: path
          schema:
            type: integer
          required: true
          example: '3'
      responses:
        '200':
          description: OK
          headers:
            Host:
              schema:
                type: string
                example: localhost:8000
            Date:
              schema:
                type: string
                example: Mon, 31 Oct 2022 08:04:17 GMT
            Connection:
              schema:
                type: string
                example: close
            X-Powered-By:
              schema:
                type: string
                example: PHP/7.4.32
            Content-Type:
              schema:
                type: string
                example: application/json
          content:
            application/json:
              schema:
                type: object
              examples:
                example-0:
                  summary: Get Task
                  value:
                    id: '10'
                    category_id: '1'
                    title: Task J
                    description: null
                    start_date: '2001-01-01'
                    finish_date: '2001-01-02'
                    status: New
                    doc_url: localhost:8000/upload/tasks/5YvFLT1WGD-Some%20File.pdf
                example-1:
                  summary: Task With Category
                  value:
                    id: '3'
                    category_id: '1'
                    title: Task C
                    description: null
                    start_date: '2001-01-01'
                    finish_date: '2001-01-02'
                    status: New
                    doc_url: localhost:8000/upload/tasks/RFX0ThCeGV-Some%20File.pdf
                    category:
                      id: '1'
                      name: Lain-lain
        '404':
          description: Not Found
          headers:
            Host:
              schema:
                type: string
                example: localhost:8000
            Date:
              schema:
                type: string
                example: Mon, 31 Oct 2022 08:04:35 GMT
            Connection:
              schema:
                type: string
                example: close
            X-Powered-By:
              schema:
                type: string
                example: PHP/7.4.32
            Content-Type:
              schema:
                type: string
                example: application/json
          content:
            application/json:
              schema:
                type: object
              example:
                message: Not Found
    put:
      tags:
        - Task
      summary: Task
      requestBody:
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                title:
                  type: string
                  example: Judul X
                start_date:
                  type: string
                  example: '2001-01-01'
                finish_date:
                  type: string
                  example: '2001-01-01'
                status:
                  type: string
                  example: New
                category_id:
                  type: integer
                  example: '3'
                doc:
                  type: string
                  format: binary
      parameters:
        - name: task_id
          in: path
          schema:
            type: integer
          required: true
          example: '16'
      responses:
        '204':
          description: No Content
          headers:
            Host:
              schema:
                type: string
                example: localhost:8000
            Date:
              schema:
                type: string
                example: Mon, 31 Oct 2022 08:31:53 GMT
            Connection:
              schema:
                type: string
                example: close
            X-Powered-By:
              schema:
                type: string
                example: PHP/7.4.32
            Content-type:
              schema:
                type: string
                example: text/html; charset=UTF-8
          content:
            text/plain:
              schema:
                type: string
              example: null
        '404':
          description: Not Found
          headers:
            Host:
              schema:
                type: string
                example: localhost:8000
            Date:
              schema:
                type: string
                example: Mon, 31 Oct 2022 08:36:09 GMT
            Connection:
              schema:
                type: string
                example: close
            X-Powered-By:
              schema:
                type: string
                example: PHP/7.4.32
            Content-Type:
              schema:
                type: string
                example: application/json
          content:
            application/json:
              schema:
                type: object
              example:
                message: Not Found
        '422':
          description: Unknown Status Code
          headers:
            Host:
              schema:
                type: string
                example: localhost:8000
            Date:
              schema:
                type: string
                example: Mon, 31 Oct 2022 08:34:19 GMT
            Connection:
              schema:
                type: string
                example: close
            X-Powered-By:
              schema:
                type: string
                example: PHP/7.4.32
            Content-Type:
              schema:
                type: string
                example: application/json
          content:
            application/json:
              schema:
                type: object
              example:
                message: Unprocessable Entity
                errors:
                  doc: You did not select a file to upload.
    delete:
      tags:
        - Task
      summary: Task
      parameters:
        - name: task_id
          in: path
          schema:
            type: integer
          required: true
          example: '13'
      responses:
        '204':
          description: No Content
          headers:
            Host:
              schema:
                type: string
                example: localhost:8000
            Date:
              schema:
                type: string
                example: Mon, 31 Oct 2022 08:35:34 GMT
            Connection:
              schema:
                type: string
                example: close
            X-Powered-By:
              schema:
                type: string
                example: PHP/7.4.32
            Content-type:
              schema:
                type: string
                example: text/html; charset=UTF-8
          content:
            text/plain:
              schema:
                type: string
              example: null
        '404':
          description: Not Found
          headers:
            Host:
              schema:
                type: string
                example: localhost:8000
            Date:
              schema:
                type: string
                example: Mon, 31 Oct 2022 08:35:53 GMT
            Connection:
              schema:
                type: string
                example: close
            X-Powered-By:
              schema:
                type: string
                example: PHP/7.4.32
            Content-Type:
              schema:
                type: string
                example: application/json
          content:
            application/json:
              schema:
                type: object
              example:
                message: Not Found
    patch:
      tags:
        - Task
      summary: Task
      requestBody:
        content:
          application/json:
            schema:
              type: object
              example:
                title: null
      parameters:
        - name: task_id
          in: path
          schema:
            type: integer
          required: true
          example: '12'
      responses:
        '204':
          description: No Content
          headers:
            Host:
              schema:
                type: string
                example: localhost:8000
            Date:
              schema:
                type: string
                example: Mon, 31 Oct 2022 08:36:36 GMT
            Connection:
              schema:
                type: string
                example: close
            X-Powered-By:
              schema:
                type: string
                example: PHP/7.4.32
            Content-type:
              schema:
                type: string
                example: text/html; charset=UTF-8
          content:
            text/plain:
              schema:
                type: string
              example: null
        '404':
          description: Not Found
          headers:
            Host:
              schema:
                type: string
                example: localhost:8000
            Date:
              schema:
                type: string
                example: Mon, 31 Oct 2022 08:37:19 GMT
            Connection:
              schema:
                type: string
                example: close
            X-Powered-By:
              schema:
                type: string
                example: PHP/7.4.32
            Content-Type:
              schema:
                type: string
                example: application/json
          content:
            application/json:
              schema:
                type: object
              example:
                message: Not Found
        '422':
          description: Unknown Status Code
          headers:
            Host:
              schema:
                type: string
                example: localhost:8000
            Date:
              schema:
                type: string
                example: Mon, 31 Oct 2022 08:37:49 GMT
            Connection:
              schema:
                type: string
                example: close
            X-Powered-By:
              schema:
                type: string
                example: PHP/7.4.32
            Content-Type:
              schema:
                type: string
                example: application/json
          content:
            application/json:
              schema:
                type: object
              example:
                message: Unprocessable Entity
                errors:
                  title: Harap masukkan judul